+++ apps/frontend/src/libs/components/app/app.tsx
@@ -12,22 +12,20 @@ import {
 	useLocation,
 } from "~/libs/hooks/hooks.js";
 import { actions as authActions } from "~/modules/auth/auth.js";
-import { actions as userActions } from "~/modules/users/users.js";
 import { QuizForm } from "~/pages/quiz/libs/components/quiz-form/quiz-form.jsx";
 
 const App: React.FC = () => {
 	const { pathname } = useLocation();
 	const dispatch = useAppDispatch();
-	const { dataStatus, users } = useAppSelector(({ users }) => ({
-		dataStatus: users.dataStatus,
-		users: users.users,
+
+	const { dataStatus } = useAppSelector(({ auth }) => ({
+		dataStatus: auth.dataStatus,
 	}));
 
 	const isRoot = pathname === AppRoute.ROOT;
 
 	useEffect(() => {
 		if (isRoot) {
-			void dispatch(userActions.loadAll());
 			void dispatch(authActions.getAuthenticatedUser());
 		}
 	}, [isRoot, dispatch]);
@@ -43,13 +41,6 @@ const App: React.FC = () => {
 			{!isLoading && isRoot && (
 				<>
 					<Header />
-					<h2>Users:</h2>
-					<h3>Status: {dataStatus}</h3>
-					<ul>
-						{users.map((user) => (
-							<li key={user.id}>{user.email}</li>
-						))}
-					</ul>
 					<QuizForm />
 				</>
 			)}