+++ apps/mobile/src/slices/auth/auth.ts
@@ -1,9 +1,10 @@
-import { getAuthenticatedUser, signUp } from "./actions";
+import { getAuthenticatedUser, signIn, signUp } from "./actions";
 import { actions } from "./auth.slice";
 
 const allActions = {
 	...actions,
 	getAuthenticatedUser,
+	signIn,
 	signUp,
 };