+++ packages/shared/src/modules/users/libs/types/user-get-all-item-response-dto.type.ts
@@ -1,5 +1,0 @@
-import { UserDto } from "./user-dto.type.js";
-
-type UserGetAllItemResponseDto = UserDto;
-
-export { type UserGetAllItemResponseDto };