+++ prettier.config.ts
@@ -1,4 +1,4 @@
-import { Config } from "prettier";
+import { type Config } from "prettier";
 
 const config: Config = {
 	arrowParens: "always",